+++ test-suite/golden-tests/enum-comparison/compare.hpp
@@ -1,0 +1,2 @@
+#pragma once
+template <typename T> T ident(T x) { return x; }